refactor/move main to pages
